requirements.txt
Django==2.2.3
psycopg2==2.8.3
djangorestframework==3.10.0
djangorestframework-jwt==1.11.0
